Moved result.py one up and set up a pass through at the multitest package level.
